Refine several todos
